solution proposed
