+++ README_CUSTOM.md
@@ -1,0 +1,5 @@
+### plugins
+
+He fet servir un plugin de "zaengle.com" seguint un article de la Rachel Opperman (https://zaengle.com/blog/error-handling-in-nuxt-apps).
+https://github.com/zaengle/error-handler.git.
+Que aprofita els "interceptors" d'axios per a desestructurar les respostes d'error, tenint en compte tant si retornen un data.error com si no.
