+++ utils/customError.js
@@ -1,0 +1,5 @@
+export const customErrors = {
+  401: `Whooops! You can't acces that.`,
+};
+
+export const defaultMessageError = `Something went wrong. Backend guy says it's frontend fault, but you and me know what the truth is`;
